Fix no extension files and add tests
