+++ roq/integration-tests/src/main/resources/application.properties
@@ -1,2 +1,4 @@
 %no-file-slugify.site.slugify-files=false
 site.escaped-pages=posts/escaped**
+%roq-and-roll.site.ignored-files=**error**
+#quarkus.log.category."io.quarkiverse.roq.generator".level=DEBUG
