+++ roq/integration-tests/src/main/resources/public/hello
@@ -1,0 +1,1 @@
+Hello
